fix(tools): ruff lint pass — REPO_ROOT/Path F821 bugs + dead code + autofixes
First ruff run surfaced a real latent bug: REPO_ROOT was never imported in 4 scripts (21c/30/30b/31) -> NameError on the final 'Wrote ...' print (reproduced on script 30). The save() calls run BEFORE that print, so the committed cache stayed valid (rerun confirmed identical numbers); the crash was cosmetic. Path likewise unimported in md_utils (lazy under future-annotations).

- F821 (10): REPO_ROOT x8 import + Path x2 (md_utils)
- F841 (12): dead vars removed; 'restraint =' prefix dropped (OpenMM addForce side-effect kept); fock/F dropped in 24b/24d (H_ab built from epsilon, not Fock)
- safe autofixes: unused imports, import-sort, empty f-strings
- behaviourally-neutral: zip(strict=False), @cache, semicolon splits
- import-smoke test guarded with noqa: F401 (intentional)

Verified: ruff 0 | in_silico 83 | ml parity 16 | firmware tables 3/3 | 0 Ruby touched

+++ tools/in_silico/lib/utils.py
@@ -15,7 +15,7 @@ def banner(msg: str) -> None:
 
 def ps_to_steps(ps: float, timestep_fs: float = 2.0) -> int:
     """Convert picoseconds to integration steps."""
-    return int(round(ps * 1000.0 / timestep_fs))
+    return round(ps * 1000.0 / timestep_fs)
 
 
 def pick_platform() -> Platform:
